Cleaning up warnings, adding name to the top of all .cpp files adding more cleanup code

+++ Asteroids V1.0/Source/BulletBag.cpp
@@ -1,3 +1,4 @@
+//Daniel Mardunovich Dama7895
 #include "BulletBag.h"
 
 BulletBag::~BulletBag()
@@ -26,7 +27,7 @@ Bullet* BulletBag::GetBullet() {
 }
 
 int BulletBag::BulletsInBag() {
-	return bullets.size();
+	return (int)bullets.size();
 }
 
 void BulletBag::Unload() {
